tech(TTB-060): Extracted all logic out of the Game component
