+++ client/src/utils/helper/gameUtils.js
@@ -100,12 +100,6 @@ export const checkIfSetUpWinningMove = (squares, x, y, icon) => {
     return numberOfWinningMovesSetUp;
 }
 
-// export const delay = (duration) => {
-//     return new Promise((resolve) => {
-//       setTimeout(() => resolve(), duration);
-//     });
-// }
-
 export const calculateSurroundingCoordinates = (x, y, includeDiagonal) => {
     let surroundingCoordinates = [];
     for (let i = -1; i <= 1; i++) {
